Add grid world engine.
Introduces an engine for grid worlds that connects with the code world config, but sets up the task in a manner more familiar to users of common grid world libraries like Lab2D or MiniGrid. It allows for easier creation of a large amount of tasks and achieves expressiveness through customizable reactive objects.

Initial examples include a four rooms task where the agent navigates across the rooms to reach the goal and a key door task. More tasks are prepared through objects like spikes that agents can step on but with a severe penalty, or doors who switches automatically between open and closed based on a formula.

PiperOrigin-RevId: 734175738
Change-Id: I0206db56c69218fa79e0a90920c8e6d90fbf93a9